Move content of StartScreen to Home and link buttons
Remove StartScreen
