+++ src/pages/public/Home.tsx
@@ -1,53 +1,69 @@
 import React from "react";
-import { Info, User, Briefcase, CheckCircle } from "react-feather";
+import { Link } from "react-router-dom";
+import { ResponsiveSvgBackground } from "~/components/Background";
+import { PrimaryButton } from "~/components/Button";
+import { ResponsiveLandingTitle } from "./components";
 
-import { Gradient } from "~/components/Gradient";
-import { Logo } from "~/components/Logo";
-import { Map } from "~/components/Map";
-import { Title } from "~/components/Title";
-import { useMobileDevice } from "~/lib/hooks";
-import { Steps } from "~/components/Steps";
+const ResetTopMarginStyles = { marginTop: 0 };
 
 type Props = {};
 
-const steps = [
-  { title: "Hinweis", Icon: Info, Content: Info },
-  { title: "Über dich", Icon: User, Content: User },
-  { title: "Konditionen", Icon: Briefcase, Content: Briefcase },
-  { title: "Fertig!", Icon: CheckCircle, Content: CheckCircle }
-];
-
 const Home = (props: Props) => {
-  const isMobile = useMobileDevice();
-  const logoSize = isMobile ? 60 : 80;
-
   return (
-    <main className="h-100vh">
-      <div className="relative z-10 bg-white flex justify-between p-2 pb-4">
-        <div className="flex items-center">
-          <Logo size={logoSize} />
-          <Title as="h1" className="text-3xl sm:text-4xl text-brand ml-1" bold>
-            FarmHelden
-          </Title>
-        </div>
-        <Gradient
-          height="15vh"
-          className="absolute left-0 top-100perc z-20"
-          block
-        />
-      </div>
-
-      <div className="fixed inset-0">
-        <div className="absolute top-0 z-20 flex items-center w-full px-2 h-full">
-          <div className="block w-full">
-            <Steps steps={steps} />
+    <>
+      <div className="relative z-10 flex flex-col h-full px-4">
+        {/* Logo and subtitle */}
+        <div className="flex-grow">
+          <div className="flex items-center h-full">
+            <ResponsiveLandingTitle
+              className="block mb-16"
+              style={ResetTopMarginStyles}
+            />
           </div>
         </div>
-        <div className="backdrop-blur fixed inset-0 z-0 h-100vh">
-          <Map hideNavigationControl hideFullScreenControl blur={4} />
+
+        {/* C2A buttons */}
+        <div className="w-full block pb-4">
+          <Link to="/map">
+            <PrimaryButton block className="tracking-wider mb-2">
+              Ich will helfen
+            </PrimaryButton>
+          </Link>
+          <Link to="/boarding-farmer">
+            <PrimaryButton block className="tracking-wider">
+              Ich benötige Hilfe
+            </PrimaryButton>
+          </Link>
+
+          <p className="mt-12 mb-2 text-center text-sm">
+            Oder hast du schon einen Account?
+          </p>
+          <PrimaryButton block theme="border" className="tracking-wider">
+            Anmelden
+          </PrimaryButton>
         </div>
       </div>
-    </main>
+
+      {/* Bottom background */}
+      <ResponsiveSvgBackground width={360} height={201}>
+        <g opacity="0.2">
+          <path
+            opacity="0.5"
+            d="M98 0C37.2 0 -58.3333 16 -109 24V322H387V16C345.333 10.6667 158.8 0 98 0Z"
+            fill="#DCB38E"
+          />
+          <path
+            d="M51 58C-27.4 58 -84.6667 79.3333 -114 95V282H364V52C349.333 62.6667 265.2 81 230 81C186 81 149 58 51 58Z"
+            fill="#94E1AE"
+          />
+          <path
+            opacity="0.5"
+            d="M108 89C29.6 89 -87.6667 110.333 -117 126V330H368C375.2 330 371 165.667 368 89C353.333 99.6667 288.2 111 253 111C209 111 206 89 108 89Z"
+            fill="#197649"
+          />
+        </g>
+      </ResponsiveSvgBackground>
+    </>
   );
 };
