+++ src/routes/index.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import { Redirect, Route, Switch } from "react-router-dom";
+import SplashScreen from "~/pages/public/SplashScreen";
 import NotFoundPage from "../pages/error/NotFound";
 import BoardingFarmer from "../pages/public/BoardingFarmer";
 import Home from "./../pages/public/Home";
 import Map from "./../pages/public/Map";
-import SplashScreen from "~/pages/public/SplashScreen";
-import StartScreen from "~/pages/public/StartScreen";
 
 export default function AppRoutes() {
   return (
@@ -20,7 +19,6 @@ export default function AppRoutes() {
 
       {/* Screens - only for presentational purposes */}
       <Route exact path="/screens/splash" component={SplashScreen} />
-      <Route exact path="/screens/start" component={StartScreen} />
 
       {/*Redirect*/}
       <Redirect
